added user id column for performing user specific todo actions

+++ internal/models/todo/delete.go
@@ -4,22 +4,23 @@ import (
 	"fmt"
 
 	"github.com/gorvk/todoapp/internal/initializers"
+	"github.com/gorvk/todoapp/internal/types"
 )
 
-func Delete(id string) error {
+func Delete(id string, userClaim types.UserClaim) error {
 	db := initializers.GetDBInstance()
 	if db == nil {
 		return fmt.Errorf("nil db instance")
 	}
 
-	stmt, err := db.Prepare("DELETE FROM Todos WHERE id = $1")
+	stmt, err := db.Prepare("DELETE FROM Todos WHERE id = $1 AND user_id = $2")
 	if err != nil {
 		return err
 	}
 
 	defer stmt.Close()
 
-	_, err = stmt.Exec(id)
+	_, err = stmt.Exec(id, userClaim.UserId)
 
 	return err
 }
